third_party/blink/web_tests/platform/generic/http/tests/media/autoplay/document-user-activation-navigation-click-expected.txt

PASS undefined is undefined.
PASS results.webAudio is "running"
PASS successfullyParsed is true

TEST COMPLETE


============== Back Forward List ==============
        http://127.0.0.1:8000/media/autoplay/document-user-activation-navigation-click.html
curr->  http://127.0.0.1:8000/media/autoplay/resources/test-autoplay.html
===============================================
